+++ src/Speculator.jl
@@ -23,12 +23,11 @@ TODO:
 =#
 
 import Base:
-    eltype, intersect, isdisjoint, isempty, issetequal,
-    issubset, iterate, length, setdiff, show, symdiff, union
+    eltype, hash, instances, intersect, issubset, iterate, length, setdiff, show, symdiff, union
 using Base:
     Threads, IdSet,
     checked_mul, isdeprecated, issingletontype, isvarargtype, loaded_modules_array,
-    mul_with_overflow, specializations, typename, uniontypes, unsorted_names, unwrap_unionall
+    mul_with_overflow, specializations, tail, typename, uniontypes, unsorted_names, unwrap_unionall
 using .Threads: @spawn
 using Core: Builtin, TypeofBottom, Typeof
 using InteractiveUtils: subtypes